+++ megalinter/linters/GroovyNpmGroovyLintLinter.py
@@ -24,9 +24,7 @@ class GroovyNpmGroovyLintLinter(Linter):
             self.try_fix = True
         # Add user-defined extra arguments if defined
         cmd += self.cli_lint_user_args
-        cmd += [
-            "--path ", dir_name,
-            "--files ", file_name]
+        cmd += ["--path ", dir_name, "--files ", f"**/{file_name}"]
         if self.config_file is not None:
             cmd += [self.cli_config_arg_name, self.config_file]
         return cmd